+++ parser/lexer/lexer.go
@@ -99,6 +99,16 @@ func (l *lexer) word() string {
 	return l.input[l.start:l.end]
 }
 
+func (l *lexer) skipComment(r rune) stateFn {
+	for l.next() != eof && l.loc.Column != 0 {
+	}
+	if r == eof {
+		l.emitEOF()
+		return nil
+	}
+	return root
+}
+
 func (l *lexer) ignore() {
 	l.start = l.end
 	l.startLoc = l.loc