+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+v1.10.0 [2020-05-22]
+-------------------
+
+#### New features
+- Adding support for inline comments with '//' 
+
+
 v1.9.0 [2020-05-12]
 -------------------
 
 #### New features
-Adding support for all bitwise operators
-Replacing all operators by constants for potential future changes
+- Adding support for all bitwise operators
+- Replacing all operators by constants for potential future changes